+++ packages/expo-ui/CHANGELOG.md
@@ -52,6 +52,7 @@
 
 ### 💡 Others
 
+- [web] Replace unmaintained `vaul` with an in-house HTML `<dialog>` bottom sheet in community and universal `BottomSheet`. (by [@kudo](https://github.com/kudo))
 - [Android] Change modifiers type and provide appContext. ([#47616](https://github.com/expo/expo/pull/47616) by [@jakex7](https://github.com/jakex7))
 - [Android] Change `Text` color props to `ColorValue`. ([#47739](https://github.com/expo/expo/pull/47739) by [@jakex7](https://github.com/jakex7))
 - [Android] Expose `getMaterialColorTokens` for `expo-widgets`. ([#48453](https://github.com/expo/expo/pull/48453) by [@jakex7](https://github.com/jakex7))